refactor:compare pathname, searchParams to determin NavItem's active state
